LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/onpaint-iframe-expected.txt


FAIL onpaint should fire after iframe changes. assert_equals: onpaint should fire after a grandchild iframe changes. expected 1 but got 0
